Initial commit for git repo of SP3_limit_algebra.py script.
This python script is being developed to perform lots of tedious algebra that
I don't want to do manually. It would have taken dozens of hours to do this by
hand, and mistakes would be almost guaranteed. This code will perform all of the
basic operations that are required to solve and simplify the expressions.