Further progress on solving the quadratic TL moment equations.
The equations can be solved easily, but there is trouble with converting all of
the floating point coefficients back to rational fractions so that they can be
displayed nicely in the LaTeX output document. Several common fractions are
defined, but as the equations are solved, the number of fractions grows very
quickly.