Changing the subterm addition and multiplication to integer math.
Instead of a floating point coefficient, the subterms now have a numerator and a
denominaotr, each of which is a list of integers, which are the prime factors of
the number. This way, it is easy to print the rational coefficients of the
subterms when we want to do so for the LaTeX output.