All bugs in SP3_limit_integer.py are fixed.
Some bugs with subterm multiplication, copying lists of terms and subterms. With
this commit, the SP3_limit_integer.py, which performs the algebra using rational
fractions instead of floating point algebra, is functional. The results agree
with the results from the SP3_limit_quadratic.py, but the floating point
numbers are automatically converted to fractions.