Fixed a bug in subterm_add for SP3_limit_integer.py.
I tried to make the subterm_add work better by simplifying the fractions
before adding them, but the method implemented was not actually fraction
addition. A find_common_factors() routine was added to the subterm class, so
that we can find common factors in the denominator of two fractions before
adding them and simplify the math.